% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbp.R
\docType{data}
\name{sbp}
\alias{sbp}
\title{Systolic Blood Pressure dataset}
\format{A data frame with 500 rows and 17 variables:
\describe{
  \item{sbp}{systolic blood pressure, in mmHg}
  \item{age}{age, in integer years}
  \item{weight}{weight, in pounds}
  \item{height}{height, in inches}
  \item{bmi}{Body Mass Index}
  \item{Gender}{sex (biological)}
  \item{Married}{marital status}
  \item{Smoke}{tobacco use}
  \item{Exercise}{exercise level}
  \item{Overweight}{categorical weight group}
  \item{Alcohol}{alcohol use}
  \item{Treatment Group}{experimental group}
  \item{Stress}{general stress level}
  \item{Salt}{salt use}
  \item{Childbearing_Potential}{ability to have children}
  \item{Income}{income level}
  \item{Education}{educational level}
}}
\source{
\url{https://ssc.ca/en/case-studies-2003-annual-meeting}
}
\usage{
sbp
}
\description{
A dataset containing physiological and socioeconomic variables for 500
Canadian patients for a blood pressure treatment study. Data from
Dr. Raymond Lam, GlaxoSmithKline, Toronto, Ontario, Canada, 2003.
}
\examples{
head(sbp)
}
\keyword{datasets}
